refactored volume slider files, removed redundant parts of names from library files
