+++ library/sliders/blur.lua
